Moved `ImageCommentsUIIntegrationTests` helpers to separate files Extracted reusable `localized(key:in:from:)` function
