+++ EssentialApp/EssentialAppTests/Helpers/FeedUIIntegrationTests+Localization.swift
@@ -8,12 +8,6 @@ import EssentialFeed
 
 extension FeedUIIntegrationTests {
 	func localized(_ key: String, file: StaticString = #filePath, line: UInt = #line) -> String {
-		let table = "Feed"
-		let bundle = Bundle(for: FeedPresenter.self)
-		let value = bundle.localizedString(forKey: key, value: nil, table: table)
-		if value == key {
-			XCTFail("Missing localized string for key: \(key) in table: \(table)", file: file, line: line)
-		}
-		return value
+		return EssentialAppTests.localized(key, in: "Feed", from: Bundle(for: FeedPresenter.self))
 	}
 }
