+++ EssentialApp/EssentialAppTests/Helpers/SharedTestHelpers.swift
@@ -20,3 +20,13 @@ func anyData() -> Data {
 func uniqueFeed() -> [FeedImage] {
 	return [FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())]
 }
+
+import XCTest
+
+func localized(_ key: String, in table: String, from bundle: Bundle, file: StaticString = #filePath, line: UInt = #line) -> String {
+	let value = bundle.localizedString(forKey: key, value: nil, table: table)
+	if value == key {
+		XCTFail("Missing localized string for key: \(key) in table: \(table)", file: file, line: line)
+	}
+	return value
+}
