+++ EssentialApp/EssentialAppTests/ImageCommentsUIIntegrationTests.swift
@@ -127,39 +127,6 @@ final class ImageCommentsUIIntegrationTests: XCTestCase {
 		
 	// MARK: - Helpers
 	
-	private func assertThat(
-		_ sut: ImageCommentsViewController,
-		isRendering imageComments: [PresentableImageComment],
-		file: StaticString = #filePath,
-		line: UInt = #line
-	) {
-		guard sut.numberOfRenderedImageComments() == imageComments.count else {
-			return XCTFail("Expected \(imageComments.count) images, got \(sut.numberOfRenderedImageComments()) instead.", file: file, line: line)
-		}
-		
-		for (index, imageComment) in imageComments.enumerated() {
-			assertThat(sut, hasViewConfiguredFor: imageComment, at: index)
-		}
-	}
-	
-	private func assertThat(
-		_ sut: ImageCommentsViewController,
-		hasViewConfiguredFor imageComment: PresentableImageComment,
-		at index: Int,
-		file: StaticString = #filePath,
-		line: UInt = #line
-	) {
-		let view = sut.imageCommentView(at: index)
-		
-		guard let cell = view as? ImageCommentCell else {
-			return XCTFail("Expected \(ImageCommentCell.self) instance, got \(String(describing: view)) instead", file: file, line: line)
-		}
-		
-		XCTAssertEqual(cell.usernameText, imageComment.author, "Expected username text to be \(String(describing: imageComment.author)) for image comment view at index \(index)", file: file, line: line)
-		XCTAssertEqual(cell.createdAtText, imageComment.createdAt, "Expected created at text to be \(String(describing: imageComment.createdAt)) for image comment view at index \(index)", file: file, line: line)
-		XCTAssertEqual(cell.commentText, imageComment.message, "Expected comment text to be \(String(describing: imageComment.author)) for image comment view at index \(index)", file: file, line: line)
-	}
-	
 	private func makeSUT(date: Date = Date(), file: StaticString = #filePath, line: UInt = #line) -> (sut: ImageCommentsViewController, loader: LoaderSpy) {
 		let loader = LoaderSpy()
 		let sut = ImageCommentsUIComposer.imageCommentsComposedWith(commentsLoader: loader, date: date)
@@ -195,84 +162,4 @@ final class ImageCommentsUIIntegrationTests: XCTestCase {
 	private func anyDate() -> Date {
 		return Date(timeIntervalSinceReferenceDate: 638556190)
 	}
-	
-	private func localized(_ key: String, file: StaticString = #filePath, line: UInt = #line) -> String {
-		let table = "ImageComments"
-		let bundle = Bundle(for: ImageCommentsPresenter.self)
-		let value = bundle.localizedString(forKey: key, value: nil, table: table)
-		if value == key {
-			XCTFail("Missing localized string for key: \(key) in table: \(table)", file: file, line: line)
-		}
-		return value
-	}
-	
-	private class LoaderSpy: ImageCommentsLoader {
-		
-		private struct TaskSpy: ImageCommentsLoaderTask {
-			let id: UUID
-			let cancelCallback: () -> Void
-			
-			func cancel() {
-				cancelCallback()
-			}
-		}
-		
-		private var imageCommentsRequests = [(ImageCommentsLoader.Result) -> Void]()
-		
-		private(set) var cancelledRequests = [UUID]()
-		
-		var loadImageCommentsCallCount: Int {
-			return imageCommentsRequests.count
-		}
-		
-		func completeCommentsLoading(with comments: [ImageComment] = [], at index: Int = 0) {
-			imageCommentsRequests[index](.success(comments))
-		}
-		
-		func completeCommentsLoadingWithError(at index: Int = 0) {
-			let error = NSError(domain: "an error", code: 0)
-			imageCommentsRequests[index](.failure(error))
-		}
-		
-		@discardableResult
-		func load(completion: @escaping (ImageCommentsLoader.Result) -> Void) -> ImageCommentsLoaderTask {
-			imageCommentsRequests.append(completion)
-			let taskId = UUID()
-			return TaskSpy(id: taskId) { [weak self] in
-				self?.cancelledRequests.append(taskId)
-			}
-		}
-	}
-}
-
-extension ImageCommentsViewController {
-	func simulateUserInitiatedImageCommentsReload() {
-		refreshControl?.simulatePullToRefresh()
-	}
-	
-	var isShowingLoadingIndicator: Bool {
-		return refreshControl?.isRefreshing == true
-	}
-	
-	func numberOfRenderedImageComments() -> Int {
-		tableView.numberOfRows(inSection: imageCommentsSection)
-	}
-	
-	func imageCommentView(at row: Int) -> UITableViewCell? {
-		let indexPath = IndexPath(row: row, section: imageCommentsSection)
-		let ds = tableView.dataSource
-		return ds?.tableView(tableView, cellForRowAt: indexPath)
-	}
-	
-	var errorMessage: String? {
-		errorView?.message
-	}
-	
-	var imageCommentsSection: Int { 0 }
-}
-
-extension ImageCommentCell {
-	var commentText: String? { commentLabel?.text }
-	var usernameText: String? { usernameLabel?.text }
-	var createdAtText: String? { createdAtLabel?.text }
 }
